requirements_runpod.txt
torch==2.0.1
accelerate==0.23.0
transformers==4.30.2
ftfy==6.1.1
albumentations==1.3.0
opencv-python==4.7.0.68
einops==0.6.0
diffusers[torch]==0.21.1
pytorch-lightning==1.9.0
bitsandbytes==0.35.0
tensorboard==2.10.1
safetensors==0.3.1
# gradio==3.16.2
altair==4.2.2
easygui==0.98.3
toml==0.10.2
voluptuous==0.13.1
# for BLIP captioning
requests
timm==0.6.12
fairscale==0.4.13
# for WD14 captioning
# tensorflow<2.11
tensorflow==2.10.1
huggingface-hub==0.15.1
google-api-core
google-api-python-client
google-cloud-storage
xformers==0.0.20
lycoris_lora
open_clip_torch==2.20.0

scikit-learn
